+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 ## Unreleased
 
 - Drop support for Laravel Lumen (#579)
+- Drop support for Laravel 5.x (#581)
+- Fix not setting the correct SDK ID and version when running the `sentry:test` command (#582)
 
 ## 2.14.0