+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Unreleased
 
+## 2.14.1
+
 - Fix not setting the correct SDK ID and version when running the `sentry:test` command (#582)
 
 ## 2.14.0